Сделал отдельный файл для класса джсон, сделал его абстрактным
